Improves stream trimming performance by approximating
Allows the stream to grow a small amount beyond the maximum backlog length until a full macro node can be removed from the radix tree, which is significantly more efficient than being precise and editing nodes. This means we will use an amount of storage equivalent to a few tens of messages more than the maximum backlog length we specify, but throughput will be higher.